+++ storage/src/main/scala/org/fiume/sketch/storage/documents/Document.scala
@@ -2,20 +2,21 @@ package org.fiume.sketch.storage.documents
 
 import cats.Eq
 import fs2.Stream
-import org.fiume.sketch.storage.documents.Model.Metadata.*
+import org.fiume.sketch.storage.documents.Document.Metadata
+import org.fiume.sketch.storage.documents.Document.Metadata.*
 
 import java.time.ZonedDateTime
 import java.util.UUID
 
-object Model:
-  case class Document[F[_]](
-    uuid: UUID,
-    metadata: Metadata,
-    bytes: Stream[F, Byte],
-    createdAt: ZonedDateTime,
-    updatedAt: ZonedDateTime
-  )
+case class Document[F[_]](
+  uuid: UUID,
+  metadata: Metadata,
+  bytes: Stream[F, Byte],
+  createdAt: ZonedDateTime,
+  updatedAt: ZonedDateTime
+)
 
+object Document:
   case class Metadata(name: Name, description: Description)
 
   object Metadata:
