+++ storage/src/main/scala/org/fiume/sketch/storage/documents/postgres/PostgresDocumentsStore.scala
@@ -9,7 +9,7 @@ import doobie.free.connection.ConnectionIO
 import doobie.implicits.*
 import doobie.postgres.implicits.*
 import fs2.Stream
-import org.fiume.sketch.storage.documents.Model.{Document, Metadata}
+import org.fiume.sketch.storage.documents.Document.Metadata
 import org.fiume.sketch.storage.documents.algebras.DocumentsStore
 import org.fiume.sketch.storage.documents.postgres.DoobieMappings
 import org.fiume.sketch.storage.documents.postgres.DoobieMappings.given
